fix: add index to logdir

+++ finetune_chatmodel_demo/scripts/finetune_ds_multiturn_8A100.sh
@@ -26,13 +26,16 @@ DATA_NAME=tool_alpaca
 DATASET_PATH=/model/chatglm3/ChatGLM3/finetune_demo/formatted_data/tool_alpaca.jsonl
 BASE_MODEL_PATH=/model/chatglm3/chatglm3-6b
 DATESTR=$(date +%Y%m%d-%H%M%S)
-OUTPUT_DIR=/model/chatglm3/checkpoints/${JOB_NAME}/${DATESTR}-${DATA_NAME}-${LR}
 
-# clean empty dir
-find /model/chatglm3/checkpoints/${JOB_NAME} -type f -name "*events.out.tfevents*" -exec rm -f {} \; || echo "clean"
-find /model/chatglm3/checkpoints/${JOB_NAME} -type d -empty -exec rmdir {} \; || echo "clean"
-find /model/chatglm3/checkpoints/${JOB_NAME} -type d -empty -exec rmdir {} \; || echo "clean"
-find /model/chatglm3/checkpoints/${JOB_NAME} -type d -empty -exec rmdir {} \; || echo "clean"
+OUTPUT_DIR=/model/chatglm3/checkpoints/${JOB_NAME}/${DATESTR}-${DATA_NAME}-${LR}/Node-$INDEX
+
+# master clean empty dir
+if [ "$INDEX" == "0" ]; then 
+  find /model/chatglm3/checkpoints/${JOB_NAME} -type f -name "*events.out.tfevents*" -exec rm -f {} \; || echo "clean"
+  find /model/chatglm3/checkpoints/${JOB_NAME} -type d -empty -exec rmdir {} \; || echo "clean"
+  find /model/chatglm3/checkpoints/${JOB_NAME} -type d -empty -exec rmdir {} \; || echo "clean"
+  find /model/chatglm3/checkpoints/${JOB_NAME} -type d -empty -exec rmdir {} \; || echo "clean"
+fi
 
 # create debug
 mkdir -p $OUTPUT_DIR
